Make completely path agnostic.
Only the last part of the path is actually needed. Making all requests
fully relative instead of root relative and stripping the start of the
path allows the entire application to run under any base path.

This can be helpful in proxy setups where this allows to just point any
path to the FileReacher without having to configure an explicit base
path in the application.